+++ HISTORY.md
@@ -12,3 +12,13 @@
 ==================
 
   * Add an option not to resave style files using injectBase64(Sync) methods
+
+0.7.0 / 2020-02-15
+==================
+
+  * Identify font type by checking magic number of the buffer using "file-type"
+  * Use absolute font paths to compare font urls in style files
+  * Add method "injectBase64.fromContent"
+  * Add method "injectBase64Sync.fromContent"
+  * Add method "injectBase64.fromBuffer"
+  * Add method "injectBase64Sync.fromBuffer"